pyproject.toml
[tool.poetry]
name = "ta"
version = "0.1.0"
description = ""
authors = ["motty.mio2 <motty.mio2@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
streamlit = "^1.11.0"
Pillow = "^9.2.0"
pandas = "^1.4.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.971"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
